refactor(workflow): move prompt memory protocol
