+++ api/core/model_runtime/memory/__init__.py
@@ -1,0 +1,3 @@
+from .prompt_message_memory import PromptMessageMemory
+
+__all__ = ["PromptMessageMemory"]
