+++ api/core/workflow/nodes/llm/protocols.py
@@ -1,10 +1,8 @@
 from __future__ import annotations
 
-from collections.abc import Sequence
 from typing import Any, Protocol
 
 from core.model_manager import ModelInstance
-from core.model_runtime.entities import PromptMessage
 
 
 class CredentialsProvider(Protocol):
@@ -20,14 +18,4 @@ class ModelFactory(Protocol):
 
     def init_model_instance(self, provider_name: str, model_name: str) -> ModelInstance:
         """Create a model instance that is ready for schema lookup and invocation."""
-        ...
-
-
-class PromptMessageMemory(Protocol):
-    """Port for loading memory as prompt messages for LLM nodes."""
-
-    def get_history_prompt_messages(
-        self, max_token_limit: int = 2000, message_limit: int | None = None
-    ) -> Sequence[PromptMessage]:
-        """Return historical prompt messages constrained by token/message limits."""
         ...
